pyproject.toml
# pyproject.toml - Modern Python Package Configuration
# Deepfake Detection Agent v2.0

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepfake-detector"
version = "2.0.0"
description = "Multi-modal deepfake detection agent with explainable AI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Deepfake Detection Team", email = "team@example.com"}
]
maintainers = [
    {name = "Deepfake Detection Team", email = "team@example.com"}
]
keywords = [
    "deepfake",
    "detection",
    "video-analysis",
    "face-detection",
    "computer-vision",
    "ai",
    "machine-learning",
    "forensics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Video",
    "Typing :: Typed"
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0,<2.0.0",
    "opencv-python>=4.5.0",
    "scipy>=1.7.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
# Development dependencies
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-asyncio>=0.20.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.10.0",
    "pre-commit>=3.0.0",
]

# Testing dependencies
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-mock>=3.10.0",
    "hypothesis>=6.50.0",
]

# Documentation dependencies
docs = [
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
    "sphinx-autodoc-typehints>=1.19.0",
]

# GPU acceleration
gpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

# Audio analysis
audio = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
]

# Research/notebook dependencies
research = [
    "jupyter>=1.0.0",
    "jupyterlab>=4.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    "pandas>=1.4.0",
    "plotly>=5.10.0",
]

# All optional dependencies
all = [
    "deepfake-detector[dev,test,docs,audio,research]",
]

[project.scripts]
deepfake-detector = "deepfake_detector.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/deepfake-detector"
Documentation = "https://deepfake-detector.readthedocs.io"
Repository = "https://github.com/your-org/deepfake-detector"
Changelog = "https://github.com/your-org/deepfake-detector/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/your-org/deepfake-detector/issues"

# =============================================================================
# Tool Configuration
# =============================================================================

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["deepfake_detector*"]

[tool.setuptools.package-data]
deepfake_detector = [
    "py.typed",
    "data/*.yaml",
    "data/*.json",
]

# -----------------------------------------------------------------------------
# pytest Configuration
# -----------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--cov=src/deepfake_detector",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=85",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "gpu: marks tests requiring GPU",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# -----------------------------------------------------------------------------
# Coverage Configuration
# -----------------------------------------------------------------------------
[tool.coverage.run]
source = ["src/deepfake_detector"]
branch = true
parallel = true
omit = [
    "*/__pycache__/*",
    "*/tests/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
precision = 2
show_missing = true

# -----------------------------------------------------------------------------
# mypy Configuration
# -----------------------------------------------------------------------------
[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_no_return = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "mediapipe.*",
    "scipy.*",
    "sklearn.*",
]
ignore_missing_imports = true

# -----------------------------------------------------------------------------
# ruff Configuration (Linting)
# -----------------------------------------------------------------------------
[tool.ruff]
target-version = "py39"
line-length = 100
fix = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "PTH",    # flake8-use-pathlib
    "PL",     # Pylint
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",    # Line too long (handled by formatter)
    "PLR0913", # Too many arguments
    "PLR2004", # Magic value in comparison
]

[tool.ruff.lint.isort]
known-first-party = ["deepfake_detector"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "PLR2004"]
"__init__.py" = ["F401"]

# -----------------------------------------------------------------------------
# black Configuration (Formatting)
# -----------------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

# -----------------------------------------------------------------------------
# isort Configuration
# -----------------------------------------------------------------------------
[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["deepfake_detector"]
skip_gitignore = true

